Refactor ReviewRequest as done in PR#109 (changes to Report are also applied)
