+++ src/main/java/com/nekromant/telegram/commands/review/ReviewCommand.java
@@ -2,7 +2,6 @@ package com.nekromant.telegram.commands.review;
 
 
 import com.nekromant.telegram.commands.MentoringReviewCommand;
-import com.nekromant.telegram.contants.CallBack;
 import com.nekromant.telegram.model.ReviewRequest;
 import com.nekromant.telegram.repository.ReviewRequestRepository;
 import com.nekromant.telegram.utils.SendMessageFactory;
@@ -13,20 +12,12 @@ import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
 import org.telegram.telegrambots.meta.bots.AbsSender;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
 import java.security.InvalidParameterException;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.util.*;
-import java.util.stream.Collectors;
 
 import static com.nekromant.telegram.contants.Command.REVIEW;
 import static com.nekromant.telegram.contants.MessageContants.*;
-import static com.nekromant.telegram.utils.FormatterUtils.defaultDateFormatter;
 
 @Slf4j
 @Component
@@ -36,6 +27,8 @@ public class ReviewCommand extends MentoringReviewCommand {
     private ReviewRequestRepository reviewRequestRepository;
     @Autowired
     private SendMessageFactory sendMessageFactory;
+    @Autowired
+    private ReviewRequestDateTimePicker reviewRequestDateTimePicker;
 
     @Autowired
     public ReviewCommand() {
@@ -51,17 +44,13 @@ public class ReviewCommand extends MentoringReviewCommand {
             String studentChatId = chat.getId().toString();
             message.setChatId(studentChatId);
 
-            ReviewRequest reviewRequest = new ReviewRequest();
             try {
                 ValidationUtils.validateArgumentsNumber(arguments);
-                reviewRequest.setStudentUserName(user.getUserName());
-                reviewRequest.setStudentChatId(studentChatId);
-                reviewRequest.setTitle(parseTitle(arguments));
-                reviewRequest.setTimeSlots(parseTimeSlots(arguments));
+                ReviewRequest reviewRequest = ReviewRequest.getTemporaryReviewRequest(user, arguments, studentChatId);
 
                 log.info("Сохранение нового реквеста {}", reviewRequest);
                 reviewRequestRepository.save(reviewRequest);
-                sendDatePicker(absSender, user.getId().toString(), reviewRequest);
+                absSender.execute(reviewRequestDateTimePicker.sendDatePicker(user.getId().toString(), reviewRequest));
             } catch (NumberFormatException e) {
                 String firstArgument = arguments[0];
                 if (firstArgument.contains("сегодня") || firstArgument.contains("завтра")) {
@@ -84,69 +73,7 @@ public class ReviewCommand extends MentoringReviewCommand {
     }
 
     private void sendAnswer(String chatId, String text, AbsSender absSender, User user) {
-        SendMessage message = new SendMessage();
-        message.setChatId(chatId);
-        message.setText(text);
+        SendMessage message = sendMessageFactory.create(chatId, text);
         execute(absSender, message, user);
-    }
-
-    private Set<Integer> parseTimeSlots(String[] strings) {
-        Set<Integer> timeSlots = new HashSet<>();
-        for (String string : strings) {
-            if (!string.toLowerCase().contains("тема")) {
-                timeSlots.add(Integer.parseInt(string));
-                if (Integer.parseInt(string) > 24 || Integer.parseInt(string) < 0) {
-                    throw new InvalidParameterException("Неверное значение часов — должно быть от 0 до 23");
-                }
-            } else {
-                return timeSlots;
-            }
-        }
-        return timeSlots;
-    }
-
-    private String parseTitle(String[] strings) {
-        for (int i = 0; i < strings.length; i++) {
-            if (strings[i].toLowerCase().contains("тема")) {
-                return Arrays.stream(strings).skip(i).collect(Collectors.joining(" "));
-            }
-        }
-        return "";
-    }
-
-    private void sendDatePicker(AbsSender absSender, String userChatId, ReviewRequest reviewRequest) throws TelegramApiException {
-        InlineKeyboardMarkup inlineKeyboardMarkup = getDatePickerInlineKeyboardMarkup(reviewRequest);
-        SendMessage message = sendMessageFactory.create(userChatId, "Выберите дату");
-        message.setReplyMarkup(inlineKeyboardMarkup);
-        absSender.execute(message);
-    }
-
-    private InlineKeyboardMarkup getDatePickerInlineKeyboardMarkup(ReviewRequest reviewRequest) {
-        InlineKeyboardMarkup inlineKeyboardMarkup = new InlineKeyboardMarkup();
-        List<List<InlineKeyboardButton>> keyboardRows = new ArrayList<>();
-
-        LocalDate currentDay = LocalDate.now(ZoneId.of("Europe/Moscow"));
-        addDateButton(keyboardRows, reviewRequest, currentDay);
-        addDateButton(keyboardRows, reviewRequest, currentDay.plusDays(1));
-        addDateButton(keyboardRows, reviewRequest, currentDay.plusDays(2));
-        addCancelButton(keyboardRows, reviewRequest);
-
-        inlineKeyboardMarkup.setKeyboard(keyboardRows);
-        return inlineKeyboardMarkup;
-    }
-
-    private void addDateButton(List<List<InlineKeyboardButton>> keyboardRows, ReviewRequest reviewRequest, LocalDate date) {
-        String dateString = date.format(defaultDateFormatter());
-        InlineKeyboardButton dateButton = new InlineKeyboardButton();
-        dateButton.setText(dateString);
-        dateButton.setCallbackData(String.join(" ", CallBack.SET_REVIEW_REQUEST_DATE_TIME.getAlias(), dateString, reviewRequest.getId().toString()));
-        keyboardRows.add(Collections.singletonList(dateButton));
-    }
-
-    private void addCancelButton(List<List<InlineKeyboardButton>> keyboardRows, ReviewRequest reviewRequest) {
-        InlineKeyboardButton cancelButton = new InlineKeyboardButton();
-        cancelButton.setText("Отмена");
-        cancelButton.setCallbackData(String.join(" ", CallBack.DENY_REVIEW_REQUEST_DATE_TIME.getAlias(), reviewRequest.getId().toString()));
-        keyboardRows.add(Collections.singletonList(cancelButton));
     }
 }
