+++ src/main/java/com/nekromant/telegram/model/ReviewRequest.java
@@ -3,11 +3,16 @@ package com.nekromant.telegram.model;
 import com.nekromant.telegram.converter.TimeSlotsConverter;
 import lombok.Data;
 import lombok.ToString;
+import org.telegram.telegrambots.meta.api.objects.User;
 
 import javax.persistence.*;
+import java.security.InvalidParameterException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Data
 @Entity
@@ -39,4 +44,38 @@ public class ReviewRequest {
 
     @Column
     private LocalDateTime bookedDateTime;
+
+    public static ReviewRequest getTemporaryReviewRequest(User user, String[] arguments, String studentChatId) {
+        ReviewRequest reviewRequest = new ReviewRequest();
+
+        reviewRequest.setStudentUserName(user.getUserName());
+        reviewRequest.setStudentChatId(studentChatId);
+        reviewRequest.setTitle(parseTitle(arguments));
+        reviewRequest.setTimeSlots(parseTimeSlots(arguments));
+        return reviewRequest;
+    }
+
+    private static Set<Integer> parseTimeSlots(String[] strings) {
+        Set<Integer> timeSlots = new HashSet<>();
+        for (String string : strings) {
+            if (!string.toLowerCase().contains("тема")) {
+                timeSlots.add(Integer.parseInt(string));
+                if (Integer.parseInt(string) > 24 || Integer.parseInt(string) < 0) {
+                    throw new InvalidParameterException("Неверное значение часов — должно быть от 0 до 23");
+                }
+            } else {
+                return timeSlots;
+            }
+        }
+        return timeSlots;
+    }
+
+    private static String parseTitle(String[] strings) {
+        for (int i = 0; i < strings.length; i++) {
+            if (strings[i].toLowerCase().contains("тема")) {
+                return Arrays.stream(strings).skip(i).collect(Collectors.joining(" "));
+            }
+        }
+        return "";
+    }
 }
